move search and survey components into separate folders
