+++ src/Components/search/Search.jsx
@@ -3,9 +3,9 @@ import React, { useEffect, useState } from "react";
 import SearchBar from "./SearchBar";
 import Results from "./Results";
 
-import data from "../data/sample.json";
+import data from "../../data/sample.json";
 
-import "./styles/search.css";
+import "../styles/search.css";
 
 const Search = (props) => {
   const [term, setTerm] = useState("");
