+++ src/Components/survey/VirtueBucketList.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import VirtueBucketListItem from "./VirtueBucketListItem";
 
-import "./styles/virtue-bucket-list.scss";
+import "../styles/virtue-bucket-list.scss";
 
 const VirtueBucketList = (props) => {
   const virtueBuckets = props.virtue_buckets.map((virtueBucket) => (
